+++ src/components/Main.js
@@ -7,6 +7,7 @@ import { useNavigate } from "react-router-dom";
 
 import Sidebar from "./Sidebar";
 import Cards from "../elements/Cards";
+import Time from "../elements/Time";
 
 const Main = () => {
   const data = useSelector((state) => state.card.post);
@@ -59,22 +60,6 @@ const Main = () => {
     GetWeight()
   }, [])
 
-  useEffect(() => {
-    interval.current = setInterval(() => {
-      setTime(new Date());
-    }, 1000);
-    return () => clearInterval(interval.current)
-  }, []);
-  const Year = time.getFullYear();
-  const Month = time.getMonth() + 1;
-  const Day = time.getDate();
-  const Hour = time.getHours();
-  const Minute = time.getMinutes();
-  const Second = time.getSeconds();
-  const TodaySecond = 86400;
-  const NowSecond = (((Hour * 60) + Minute) * 60) + Second
-  const TodayPer = ((NowSecond / TodaySecond) * 100).toFixed(1)
-
   const dispatch = useDispatch();
   const navigate = useNavigate();
 
@@ -86,15 +71,7 @@ const Main = () => {
           <Item1>
             <CircleWrap>
               <InsideCircle>
-                <TopText>
-                  <p>{Year}년 {Month < 10 ? '0' + Month : Month}월 {Day < 10 ? '0' + Day : Day}일</p>
-                  <p>{Hour < 10 ? '0' + Hour : Hour} : {Minute < 10 ? '0' + Minute : Minute} : {Second < 10 ? '0' + Second : Second}</p>
-                  <p>지금은 단식시간 입니다.</p>
-                </TopText>
-                <BottomText>
-                  <p>단식 시간 : 13시 ~ 08시</p>
-                  <p>음식 섭취 가능 시간 : 08시 ~ 13시</p>
-                </BottomText>
+                <Time time={time} setTime={setTime} />
               </InsideCircle>
             </CircleWrap>
           </Item1>
@@ -233,15 +210,6 @@ font-weight: 700;
 font-size: 16px;
 line-height: 14px;
 `
-
-const TopText = styled.div`
-  color: #ffffff;
-  font-size: 24px;
-  text-shadow: -1px 0 #FE7770, 0 1px #FE7770, 1px 0 #FE7770, 0 -1px #FE7770;
-`;
-
-const BottomText = styled.div`
-`;
 
 const CardList = styled.div`
   width: 100%;
